Review comments - 2
