+++ server/constants.js
@@ -1,0 +1,5 @@
+const constants = {
+    HINT_WAIT_TIME: 25 * 1000,
+};
+
+module.exports = constants;
